feat: merge unauthorized with authorized
