+++ DiscordBotsList.Api/DiscordBotListApi.cs
@@ -2,8 +2,12 @@
 using DiscordBotsList.Api.Internal.Queries;
 using DiscordBotsList.Api.Objects;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Net.Http.Json;
+using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
 
@@ -11,13 +15,18 @@ namespace DiscordBotsList.Api
 {
     public class DiscordBotListApi
     {
-        protected const string baseEndpoint = "https://top.gg/api/";
+        private const string baseEndpoint = "https://top.gg/api/";
         private readonly JsonSerializerOptions _serializerOptions;
-        protected HttpClient _httpClient;
+        private readonly ulong _selfId;
+        private readonly HttpClient _httpClient;
 
-        public DiscordBotListApi()
+        public DiscordBotListApi(ulong selfId, string token)
         {
+            _selfId = selfId;
+
             _httpClient = new HttpClient();
+            _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+
             _serializerOptions = new JsonSerializerOptions();
             _serializerOptions.Converters.Add(new ULongToStringConverter());
         }
@@ -35,7 +44,7 @@ namespace DiscordBotsList.Api
             {
                 throw new ArgumentException("sort must be 'monthlyPoints', 'id', or 'date'.");
             }
-            
+
             if (count < 0 || count > 500)
             {
                 throw new ArgumentOutOfRangeException(nameof(count), "count mustn't be negative or exceed 500.");
@@ -58,7 +67,7 @@ namespace DiscordBotsList.Api
         /// <returns>Bot Object</returns>
         public async Task<IDblBot> GetBotAsync(ulong id)
         {
-            return (IDblBot)await GetBotAsync<Bot>(id);
+            return await GetBotAsync<Bot>(id);
         }
 
         /// <summary>
@@ -91,7 +100,7 @@ namespace DiscordBotsList.Api
         /// <typeparam name="T">Type to parse to</typeparam>
         /// <param name="url">Url to get from</param>
         /// <returns>Object of type T</returns>
-        protected async Task<T> GetAsync<T>(string url)
+        private async Task<T> GetAsync<T>(string url)
         {
             var t = await _httpClient.GetAsync(baseEndpoint + url);
             var result = t.IsSuccessStatusCode
@@ -107,6 +116,57 @@ namespace DiscordBotsList.Api
         public async Task<bool> IsWeekendAsync()
         {
             return (await GetAsync<WeekendObject>("weekend")).Weekend;
+        }
+
+        
+        /// <summary>
+        ///     Gets your own bot with as an ISelfBot
+        /// </summary>
+        /// <returns>your own bot with as an ISelfBot</returns>
+        public async Task<IDblSelfBot> GetMeAsync()
+        {
+            var bot = await GetBotAsync<SelfBot>(_selfId);
+            bot.api = this;
+            return bot;
+        }
+
+        /// <summary>
+        ///     Gets unique voters that have voted on your bot
+        ///     Max 1000, If you have more, you MUST use WEBHOOKS instead.
+        /// </summary>
+        /// <param name="page">The page number, defaults to 1</param>
+        /// <returns>A list of voters</returns>
+        public async Task<List<IDblEntity>> GetVotersAsync(int page = 1)
+        {
+            return (await GetAsync<List<Entity>>(Utils.CreateQuery($"bots/{_selfId}/votes?page={Math.Max(page, 1)}"))).Cast<IDblEntity>().ToList();
+        }
+
+        /// <summary>
+        ///     Update your stats
+        /// </summary>
+        /// <param name="guildCount">count of guilds</param>
+        public async Task UpdateStatsAsync(int guildCount)
+        {
+            if (guildCount <= 0)
+            {
+                throw new ArgumentOutOfRangeException(nameof(guildCount), "guildCount cannot be less than 1.");
+            }
+
+            var json = JsonSerializer.Serialize(new GuildCountObject(guildCount));
+            var httpContent = new StringContent(json, Encoding.UTF8, "application/json");
+
+            await _httpClient.PostAsync($"{baseEndpoint}/bots/stats", httpContent);
+        }
+
+        /// <summary>
+        ///     returns true if user have voted for the past 12 hours
+        /// </summary>
+        /// <param name="userId">Amount of days to filter</param>
+        /// <returns>True or False</returns>
+        public async Task<bool> HasVoted(ulong userId)
+        {
+            var url = $"bots/check?userId={userId}";
+            return (await GetAsync<HasVotedObject>(url)).HasVoted.GetValueOrDefault(0) == 1;
         }
     }
 }
